#47 extract interfaces into own file
